spread stuff into bunch of files
