+++ katwikOpsys.h
@@ -6,9 +6,6 @@
 #ifndef KATWIK_HEADER
 #define KATWIK_HEADER
 
-// lol
-#define HEADER_OKAY
-
 #ifndef __GNUC__
 #error bruhhhh moment no GNU?
 #undef HEADER_OKAY
@@ -27,14 +24,11 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
+//#include <string.h>
 #include <sys/socket.h>
 #include <sys/wait.h>
 #include <time.h>
 #include <unistd.h>
-
-// for printing sum noic stack traces
-#include <execinfo.h>
 
 // this one can actually be removed b/c
 // no actual code requires it,
@@ -42,166 +36,9 @@
 // so yeahh I guess this makes life easier sometimes?
 #include <mqueue.h>
 
-#ifndef ERR_MULTIPROCESS
-#pragma GCC error "ERR_MULTIPROCESS macro undefined.\nDefine it above your #include for this file.\nFor multi-process programs:\n#define ERR_MULTIPROCESS 1\nFor single-process programs:\n#define ERR_MULTIPROCESS 0"
-#undef HEADER_OKAY
-#elif ERR_MULTIPROCESS != 1 && ERR_MULTIPROCESS != 0
-#pragma GCC error "ERR_MULTIPROCESS invalid value\nFor multi-process programs:\n#define ERR_MULTIPROCESS 1\nFor single-process programs:\n#define ERR_MULTIPROCESS 0"
-#undef HEADER_OKAY
-#endif // ERR_MULTIPROCESS
+// settings n shiet for le header
+#include "options.h"
 
-// String for usage error,
-// SHOULD be manually defined above include for this file,
-// if USAGE macro is going to be used, otherwise define as blank
-// string or smthn
-#ifndef USAGE_STRING
-#pragma GCC warning "USAGE_STRING macro undefined.\nDefine it above your #include for this file.\nEg:\n#define USAGE_STRING \"arg1 arg2 etc\""
-#endif // ERR_MULTIPROCESS
-
-// For less dumb header noise at compile time lul:
-#ifndef MUNDANE_MESSAGES
-#define MUNDANE_MESSAGES 1
-#elif MUNDANE_MESSAGES == 1
-#pragma message "Mundane messages have been manually turned on"
-#elif MUNDANE_MESSAGES != 0
-#pragma GCC error "MUNDANE_MESSAGES macro bad values.\nValid values are 0 (off) or 1 (on)"
-#undef HEADER_OKAY
-#endif // MUNDANE_MESSAGES
-
-// 1 for some fun 4bit colors, 0 for boring
-#ifndef ANSI_COLOR_FIESTA
-#define ANSI_COLOR_FIESTA 1
-#elif ANSI_COLOR_FIESTA == 1
-#if MUNDANE_MESSAGES
-#pragma message "Noice colorz have been manually turned on"
-#endif // MUNDANE_MESSAGES
-#elif ANSI_COLOR_FIESTA == 0
-#pragma GCC warning "Bruhh no noice colorz?"
-#else
-#pragma GCC error "ANSI_COLOR_FIESTA macro bad value.\nValid values are 0 (off) or 1 (on)"
-#undef HEADER_OKAY
-#endif // ANSI_COLOR_FIESTA
-
-// 1 to use rand_r in randomizers
-#ifndef USE_RAND_R
-#define USE_RAND_R 0
-#elif USE_RAND_R == 1
-#if MUNDANE_MESSAGES
-#pragma message "Usage of rand_r has been manully turned on, you will have to pass a seed to all randomizers"
-#endif // MUNDANE_MESSAGES
-#elif USE_RAND_R == 0
-#pragma message "Usage of rand_r has been manully turned off"
-#else
-#pragma GCC error "USE_RAND_R macro bad value.\nValid values are 0 (off) or 1 (on)"
-#undef HEADER_OKAY
-#endif // USE_RAND_R
-
-// 1 for moar info, 0 normally
-#ifndef DEBUG_MODE
-#define DEBUG_MODE 0
-#elif DEBUG_MODE == 1
-#if MUNDANE_MESSAGES
-#pragma message "Debug mode has been manually turned on"
-#endif // MUNDANE_MESSAGES
-#elif DEBUG_MODE == 0
-#if MUNDANE_MESSAGES
-#pragma message "Debug mode has been manually turned off"
-#endif // MUNDANE_MESSAGES
-#elif DEBUG_MODE == 0
-#else
-#pragma GCC error "DEBUG_MODE macro bad value.\nValid values are 0 (off) or 1 (on)"
-#undef HEADER_OKAY
-#endif // DEBUG_MODE
-
-// seems dumb but might want to also have the option to disable these
-// even when __GNUC__ is defined 
-#ifndef ATTRIBUTES_ALLOWED
-#define ATTRIBUTES_ALLOWED 1
-#elif ATTRIBUTES_ALLOWED == 1
-#if MUNDANE_MESSAGES
-#pragma message "Attributes have been manually allowed"
-#endif // MUNDANE_MESSAGES
-#elif ATTRIBUTES_ALLOWED == 0
-#if MUNDANE_MESSAGES
-#pragma message "Attributes have been manually disallowed"
-#endif // MUNDANE_MESSAGES
-#else
-#pragma GCC error "ATTRIBUTES_ALLOWED macro bad value.\nValid values are 1 (allowed) or 0 (disallowed)"
-#undef HEADER_OKAY
-#endif // ATTRIBUTES_ALLOWED
-
-// config for whether or not some int-error functions from this header should just be void,
-// immediately using ERR instead of returning an error code
-#ifndef RETURN_ERR
-#define RETURN_ERR 0
-#elif RETURN_ERR == 1
-#if MUNDANE_MESSAGES
-#pragma message "Returning error codes from this header has been manually allowed"
-#endif // MUNDANE_MESSAGES
-#elif RETURN_ERR == 0
-#if MUNDANE_MESSAGES
-#pragma message "Returning error codes from this header has been manually disallowed"
-#endif // MUNDANE_MESSAGES
-#else
-#pragma GCC error "RETURN_ERR macro bad value.\nValid values are 1 (allowed) or 0 (disallowed)"
-#undef HEADER_OKAY
-#endif // RETURN_ERR
-
-// mainly for braced groups in the seksi macros lol
-#ifndef EXTENSIONS_ALLOWED
-#define EXTENSIONS_ALLOWED 1
-#elif EXTENSIONS_ALLOWED == 1
-#if MUNDANE_MESSAGES
-#pragma message "header's usage of __extension__ has been manually allowed"
-#endif // MUNDANE_MESSAGES
-#elif EXTENSIONS_ALLOWED == 0
-#if MUNDANE_MESSAGES
-#pragma message "header's usage of __extension__ has been manually disallowed"
-#endif // MUNDANE_MESSAGES
-#else
-#pragma GCC error "EXTENSIONS_ALLOWED macro bad value.\nValid values are 1 (allowed) or 0 (disallowed)"
-#undef HEADER_OKAY
-#endif // EXTENSIONS_ALLOWED
-
-// lol these List methods are just a slightly upgraded version from
-// my Sem1 early exam prep
-#ifndef LIST_TYPE
-#if MUNDANE_MESSAGES
-//#pragma message "LIST_TYPE undefined, defaulting to int"
-#endif // MUNDANE_MESSAGES
-#define LIST_TYPE int
-#endif // LIST_TYPE
-
-#ifndef TRACE_ON_ERR
-#define TRACE_ON_ERR 1
-#elif TRACE_ON_ERR == 1
-#if MUNDANE_MESSAGES
-#pragma message "Tracing when ERR happens has been manually enabled"
-#endif // MUNDANE_MESSAGES
-#elif TRACE_ON_ERR == 0
-#if MUNDANE_MESSAGES
-#pragma message "Tracing when ERR happens has been manually disallowed"
-#endif // MUNDANE_MESSAGES
-#else
-#pragma GCC error "TRACE_ON_ERR macro bad value.\nValid values are 1 (allowed) or 0 (disallowed)"
-#undef HEADER_OKAY
-#endif // TRACE_ON_ERR
-
-#ifndef EXEC_ALLOWED
-#define EXEC_ALLOWED 1
-#elif EXEC_ALLOWED == 1
-#if MUNDANE_MESSAGES
-#pragma message "Header's usage of exec (and similar functions) has been manually allowed"
-#endif // MUNDANE_MESSAGES
-#elif EXEC_ALLOWED == 0
-#if MUNDANE_MESSAGES
-#pragma message "Header's usage of exec (and similar functions) has been manually disallowed"
-#endif // MUNDANE_MESSAGES
-#else
-#pragma GCC error "EXEC_ALLOWED macro bad value.\nValid values are 1 (allowed) or 0 (disallowed)"
-#undef HEADER_OKAY
-#endif // EXEC_ALLOWED
- 
 // rest of header will now be invisible after the preprocesser if the first section wasn't set
 // up properly, to reduce the flooding of errors from the remaining section being invalid as a consequence.
 // in some cases this will in exchange increase "Wimplicit-function-declaration" and dumb linker errors,
@@ -210,427 +47,21 @@
 #ifndef HEADER_OKAY
 #pragma message "Some macro wasn't set up properly, rest of header won't be preprocessed"
 #else
+
+// cute colorz
+#include "colors.h"
+
+// extra attributes
+#include "extraAttr.h"
+
+// smol things
+#include "littleHelpers.h"
+
+// error stuff
+#include "errorHelpers.h"
+
 // long story short, block comments don't nest nicely lol
 #define TEMP_REMOVE 0
-
-// moi attribyut
-#if ATTRIBUTES_ALLOWED
-#define NON_ATTR_IF(expr) if (false)
-#define MY_WARN_UNUSED __attribute__((warn_unused_result))
-#define MY_NON_NULL(...) __attribute__(( nonnull(__VA_ARGS__) ))
-#define MY_DEPRECATE(msg) __attribute__ (( deprecated (msg) ))
-#define MY_FORMAT(...) __attribute__(( format (__VA_ARGS__) ))
-#define MY_INLINE __attribute__((always_inline))
-#else // ATTRIBUTUES_ALLOWED
-#define NON_ATTR_IF(expr) if (expr) /* for some parameter checks without getting Wnonnull-compare */
-#define MY_WARN_UNUSED
-#define MY_NON_NULL(...)
-#define MY_DEPRECATE(msg)
-#define MY_FORMAT(...)
-#define MY_INLINE
-#endif // ATTRIBUTUES_ALLOWED
-
-// big number
-#define GIGA 1000000000L
-
-#if ANSI_COLOR_FIESTA
-#define RESET_ESC		"\033[0m"
-#define DBLUE RESET_ESC		"\033[34m"
-#define RED_BLD_INV RESET_ESC	"\033[31;1;7m"
-#define LBLUE RESET_ESC		"\033[96m"
-#define LRED_BLD RESET_ESC	"\033[91;1m"
-#define LRED_BLD_INV RESET_ESC	"\033[91;1;7m"
-#define VIOLET_BLD RESET_ESC	"\033[35;1m"
-#define BWHITE RESET_ESC 	"\033[97m"
-#define BWHITE_INV RESET_ESC	"\033[97;7m"
-#define GREEN_BLD RESET_ESC	"\033[32;1m"
-#define YELLLOW RESET_ESC 	"\033[33m"
-
-#else
-#define RESET_ESC		""
-#define DBLUE			""
-#define RED_BLD_INV		""
-#define LBLUE			""
-#define LRED_BLD		""
-#define LRED_BLD_INV		""
-#define VIOLET_BLD		""
-#define BWHITE			""
-#define BWHITE_INV		""
-#define GREEN_BLD		""
-#define YELLLOW			""
-#endif // ANSI_COLOR_FIESTA
-
-//  requires:
-//  __FILE__, __LINE__, __func__, pid, ppid, tid
-#define INFO_FORMAT GREEN_BLD"%s" RESET_ESC":" YELLLOW"%d"\
-		RESET_ESC" in " VIOLET_BLD"%s" RESET_ESC", PID:"\
-		BWHITE"%06ld" RESET_ESC" PPID:" BWHITE"%06ld"\
-		RESET_ESC" TID:" BWHITE"%ld" RESET_ESC
-
-// NB: isMultiprocess is known at compile time
-// so any regular compiler won't even assemble
-// the if statement
-#define ERR_EXIT(isMultiprocess) do {\
-	if (isMultiprocess) kill(0, SIGKILL); /* Process Group */\
-	else exit(EXIT_FAILURE); /* Own Process Only */\
-} while(0)
-
-// dodges ASAN sometimes :)
-// a similar macro should be in safe-malloc.h but I can't include that for some reason :(
-#define FREE(ptr) do {\
-	free(ptr);\
-	(ptr) = NULL;\
-} while(0)
-	
-// goodbye Wunused-parameter/Wunused-variable lol
-#define UNUSED(thing) do {\
-	(void) (thing);\
-} while(0)
-
-// useful to force stringify some stuff
-// 2 layers because uhhhh umm
-// something something stackoverflow
-#define TO_STR_(s) #s
-#define TO_STR(X) TO_STR_(X)
-
-#if EXEC_ALLOWED
-#define MAX_EXEC_STR 1000
-#define MAX_EXEC_LINE 200
-#define EXEC_ONLY if (1)
-#else // EXEC_ALLOWED
-#define MAX_EXEC_STR 0
-#define MAX_EXEC_LINE 0
-#define EXEC_ONLY if (0)
-#endif // EXEC_ALLOWED
-
-// lmao bruh momnt
-static inline ssize_t MY_INLINE MY_NON_NULL(1)
-	charPos(const char* s, int c) {
-		char* loc = strchr(s, c);
-		return loc ? loc - s : -1;
-	}
-
-// cutely prints output of "addr2line -f -e filename address" cmd to a stream
-static void MY_NON_NULL(1, 2, 3)
-	myAddr2LinePrinter(FILE* stream, char* restrict line1, char* restrict line2,
-			int oflag
-			// possible options for oflag bitfield
-#define TRUNCATE_RESERVED 01
-#define TRUNCATE_TRACE_FUNCTIONS 02
-#define TRUNCATE_UNKNOWN 04
-#define TRUNCATE_ALOT TRUNCATE_RESERVED | TRUNCATE_TRACE_FUNCTIONS | TRUNCATE_UNKNOWN
-			) {
-		if (oflag & TRUNCATE_RESERVED && *line1 == '_') {
-			fprintf(stream, VIOLET_BLD"\t[" BWHITE" Reserved Symbol " VIOLET_BLD"]\n"RESET_ESC);
-			return;
-		} else if (oflag & TRUNCATE_UNKNOWN && *line1 == '?') {
-			fprintf(stream, VIOLET_BLD"\t[" BWHITE" ?? " VIOLET_BLD"]\n"RESET_ESC);
-			return;
-		} else if (oflag & TRUNCATE_TRACE_FUNCTIONS && ( !strcmp(line1, "myStackTracer\n") || !strcmp(line1, "backtrace\n"))) {
-			fprintf(stream, VIOLET_BLD"\t[" BWHITE" Stack Trace Function " VIOLET_BLD"]\n"RESET_ESC);
-			return;
-		}
-
-		fprintf(stream, VIOLET_BLD"\t");
-		// lmfao wtf?? print the ling char by char just to ignore the '\n'??
-		// lmao make this better
-		for (char* curr = line1; *curr && *curr != '\n'; ++curr) {
-			fputc(*curr, stream);
-		}
-
-		fprintf(stream, BWHITE" ( )\n" GREEN_BLD"\t");
-		int lineNumberPrinted = 0;
-		for (char* curr = line2; *curr; ++curr) {
-			switch (*curr) {
-				case ':':
-					{
-						fprintf(stream, RESET_ESC":" YELLLOW);
-						lineNumberPrinted = 1; // i guess lol, usually works anyway
-						// just uhh don't have ':' in paths tho? idk lol
-						break;
-					}
-				case '(':
-					{
-						if (lineNumberPrinted) {
-							// for printing discriminator thingy?
-							fprintf(stream, RESET_ESC"(");
-						} else goto justPutCharLol; // the usual -Wimplicit-fallthrough hax
-						break;
-					}
-				default:
-justPutCharLol:
-					{
-						fputc(*curr, stream);
-						break;
-					}
-			}
-		}
-
-		fprintf(stream, RESET_ESC);
-		return;
-	}
-
-// NB: DON'T link with -rdynamic b/c of this function
-// the internet tells lies lol
-//
-// also this function is pretty unsafe lol
-// you can turn it off with #define TRACE_ON_ERR 0
-#define MAX_TRACE 150 /* even super complicated programs don't need this much tbh lol */
-void myStackTracer(FILE* stream,
-		int oflag // passed to myAddr2LinePrinter
-		) {
-	void* traceBuffer[MAX_TRACE];
-	int size = backtrace(traceBuffer, MAX_TRACE);
-	char** traceStrings;
-	fprintf(stream, RESET_ESC"Call stack trace");
-
-	if (NULL == (traceStrings = backtrace_symbols(traceBuffer, size))) {
-		fprintf(stream, LRED_BLD" [Couldn't backtrace_symbols( )]\n"RESET_ESC);
-		FREE(traceStrings);
-		return;
-	}
-
-	FILE* execToFile;
-	int ranOutOfSpaceAt;
-	EXEC_ONLY {
-		// extra char for '\0'
-		char execBuffer[MAX_EXEC_STR + 1] = {0};
-
-		// create the execution string for addr2line
-		char* s2Curr = execBuffer;
-
-		char* commandStr = "addr2line -f -e ";
-		size_t commandStrSz = strlen(commandStr);
-		ranOutOfSpaceAt = size;
-
-		// TODO: better variable names here
-		// ALSO BRUH WTF WHY DIDN'T I JUST USE MEMCPY AAAAA
-		for (int i = 0, fail = 0; i < size; ++i) {
-			char* sCurr = traceStrings[i]; 
-			//s2Curr += sprintf(s2Curr, "addr2line -f -e ");
-
-			if (s2Curr > execBuffer + MAX_EXEC_STR - commandStrSz)  {
-				ranOutOfSpaceAt = i;
-				//printf("SPRINTF\n");
-				break;
-			} else {
-				memcpy(s2Curr, commandStr, commandStrSz);
-				s2Curr += commandStrSz;
-			}
-
-			for (int done = 0; *sCurr && !done; ++sCurr, ++s2Curr) {
-				if (s2Curr > execBuffer + MAX_EXEC_STR - 1)  {
-					ranOutOfSpaceAt = i;
-					//printf("SINGLE\n");
-					fail = 1;
-					break;
-				} else
-					switch (*sCurr) {
-						case ')':
-							{
-								done = 1;
-								goto replaceBrace; // Wimplicit-fallthrough hax lol
-							}
-						case '(':
-replaceBrace:
-							{
-								*s2Curr = ' ';
-								break;   
-							}
-						default:
-							{
-								*s2Curr = *sCurr;
-								break;
-							}
-					}
-			}
-			if (fail) {
-				break;
-			}
-
-			if (s2Curr > execBuffer + MAX_EXEC_STR - 1)  {
-				ranOutOfSpaceAt = i;
-				printf("SINGLE2\n");
-				break;
-			} else {
-				*s2Curr = ';';
-				++s2Curr;
-			}
-		}
-
-		//fprintf(stderr, "%s\n%d\n\n", execBuffer, ranOutOfSpaceAt);
-		if (NULL == (execToFile = popen(execBuffer, "r"))) {
-			fprintf(stream, LRED_BLD" [Couldn't popen( )]\n"RESET_ESC);
-			FREE(traceStrings);
-			return;
-		}
-	} else {
-		UNUSED(execToFile); // necessary? idk lol
-		UNUSED(ranOutOfSpaceAt);
-	}
-
-	fprintf(stream,
-	", Top %d frames from top to bottom:\n"
-	"god help you if this trace happens in multiple threads at once lol " BWHITE"\u2620" RESET_ESC"\n"
-	, size);
-
-	EXEC_ONLY {
-		fprintf(stream, "Also, line numbers in this trace are SOMEtimes slightly inaccurate for now, idk whyyy X|\n");
-	}
-
-	char line1[MAX_EXEC_LINE + 1], line2[MAX_EXEC_LINE + 1];
-	for (int i = 0, badTraceMessagePrinted = 0; i < size; ++i) {
-		fprintf(stream, BWHITE"#%d\t" RESET_ESC" %s\n"RESET_ESC, i, traceStrings[i]);
-
-		EXEC_ONLY {
-			if (ranOutOfSpaceAt == i) {
-				if (!badTraceMessagePrinted) {
-					fprintf(stream, LRED_BLD
-							"\t[ Ran out of stack space for buffer to pretty-fy stack trace ]\n"
-							"\t[ Consider setting MAX_EXEC_STR to a higher value ]\n"RESET_ESC);
-					badTraceMessagePrinted = 1;
-				}
-				continue;
-			}
-
-			// extra char for '\0'
-			fgets(line1, MAX_EXEC_LINE, execToFile);
-			fgets(line2, MAX_EXEC_LINE, execToFile);
-			myAddr2LinePrinter(stream, line1, line2, oflag);
-		}
-	}
-
-	EXEC_ONLY {
-		if (pclose(execToFile) == -1) {
-			fprintf(stream, LRED_BLD"[Couldn't pclose( )]\n"RESET_ESC);
-			perror(LRED_BLD"[Couldn't pclose( )]"RESET_ESC);
-		}
-	}
-
-	fflush(stream);
-	FREE(traceStrings);
-}
-
-#if TRACE_ON_ERR
-#define ERR_STACK_TRACE() myStackTracer(stderr, TRUNCATE_ALOT); fprintf(stderr, "\n")
-#else // TRACE_ON_ERR
-#define ERR_STACK_TRACE()
-#endif // TRACE_ON_ERR
-
-// NB: the manuals state that getpid/getppid are always successful and no errors are reserved
-// Also this version can run some optional extra code (usually cleanup) before exiting
-#define ERR_(source, cleanup) do {\
-	int originalErrno = errno;/* just in case something weird happens lmao */\
-	\
-	close(STDOUT_FILENO); /* sometimes some weeeird stuff happens without this lol*/\
-	/* especially in multithreaded environments */\
-	\
-	fprintf(stderr,\
-			RED_BLD_INV"errno=%d" RESET_ESC" " INFO_FORMAT "\n"\
-			, errno, __FILE__, __LINE__, __func__,\
-			(long) getpid(), (long) getppid(), (long) pthread_self());\
-	ERR_STACK_TRACE();\
-	errno = originalErrno;\
-	perror(source);\
-	cleanup\
-	ERR_EXIT(ERR_MULTIPROCESS);\
-} while(0)
-// 700IQ setup for having macros with optional arguments
-// without simply using __VAR_ARGS__ lol
-// courtesy of stackoverflow.com/q/3046889
-#define ERR_0_ARGS() ERR_(LBLUE"Error Description"RESET_ESC,)
-#define ERR_1_ARG(source) ERR_(source,)
-#define ERR_2_ARGS(source, cleanup) ERR_(source, cleanup)
-#define ERR_X_ARGS(x, A, B, FUNC, ...) FUNC
-#define ERR(...) ERR_X_ARGS(, ##__VA_ARGS__,\
-	ERR_2_ARGS(__VA_ARGS__),\
-	ERR_1_ARG(__VA_ARGS__),\
-	ERR_0_ARGS(__VA_ARGS__))\
-
-#if EXTENSIONS_ALLOWED
-
-// kinda overkill? but I guess this version should let you use the
-// comparison value afterwards
-#define USAGE(correctUsage) (__extension__({\
-	bool assertion = (bool) (correctUsage);\
-	if (!assertion) {\
-		ERR_NEG(fprintf(stderr,\
-					RED_BLD_INV"BAD ARGS" RESET_ESC": " LRED_BLD"!" RESET_ESC"( " LBLUE#correctUsage\
-					RESET_ESC" )\n\n" LRED_BLD"Usage" RESET_ESC":\n" BWHITE"$ %s" RESET_ESC" %s\n"\
-					, argv[0], USAGE_STRING));\
-		ERR_EXIT(ERR_MULTIPROCESS);\
-	}\
-	assertion;\
-}))
-
-#define ERR_IF(expr, condition, castType, ...) (__extension__({\
-	errno = 0;\
-	castType exprVal = (castType) (expr);\
-	if ( condition exprVal )\
-	{ERR(LRED_BLD#condition RESET_ESC" ( " LBLUE#expr RESET_ESC" ) " RESET_ESC, ##__VA_ARGS__);}\
-	exprVal;\
-}))
-#define DBGprintf(format, ...) (__extension__({\
-	int printfRet = ERR_NEG(printf(LRED_BLD_INV"DBG:" RESET_ESC" " INFO_FORMAT "\n" BWHITE format RESET_ESC\
-				, __FILE__, __LINE__, __func__,\
-				(long) getpid(), (long) getppid(), (long) pthread_self(), ##__VA_ARGS__));\
-	ERR_EOF(fflush(stdout));\
-	printfRet;\
-}))
-
-#else // EXTENSIONS_ALLOWED
-
-#define USAGE(correctUsage) do {\
-	if (!(correctUsage)) {\
-		ERR_NEG(fprintf(stderr,\
-					RED_BLD_INV"BAD ARGS" RESET_ESC": " LRED_BLD"!" RESET_ESC"( " LBLUE#correctUsage\
-					RESET_ESC" )\n\n" LRED_BLD"Usage" RESET_ESC":\n" BWHITE"$ %s" RESET_ESC" %s\n"\
-					, argv[0], USAGE_STRING));\
-		ERR_EXIT(ERR_MULTIPROCESS);\
-	}\
-} while(0)
-
-// NB: castType is ignored here, because I'm too lazy lol
-// not the best solution but whatever
-// TODO: FIX THIS LOL
-#define ERR_IF(expr, condition, castType, ...) do {\
-	errno = 0;\
-	castType exprVal = (castType) (expr);\
-	if ( condition exprVal )\
-	{ERR(LRED_BLD#condition RESET_ESC" ( " LBLUE#expr RESET_ESC" ) " RESET_ESC, ##__VA_ARGS__);}\
-} while (0)
-#define DBGprintf(format, ...) do {\
-	ERR_NEG(printf(LRED_BLD_INV"DBG:" RESET_ESC" " INFO_FORMAT "\n" BWHITE format RESET_ESC\
-				, __FILE__, __LINE__, __func__,\
-				(long) getpid(), (long) getppid(), (long) pthread_self(), ##__VA_ARGS__));\
-	ERR_EOF(fflush(stdout));\
-} while(0)
-
-#endif // EXTENSIONS_ALLOWED
-
-#define ERR_NEG1(expr, ...) ERR_IF(expr, -1 ==, int, ##__VA_ARGS__)
-#define ERR_NEG(expr, ...) ERR_IF(expr, 0 > , int, ##__VA_ARGS__)
-#define ERR_NON_ZERO(expr, ...) ERR_IF(expr, 0 !=, int, ##__VA_ARGS__)
-#define ERR_EOF(expr, ...) ERR_IF(expr, EOF ==, int, ##__VA_ARGS__)
-#define ERR_NULL(expr, ...) ERR_IF(expr, NULL ==, void*, ##__VA_ARGS__)
-#define ERR_SSIZE_T(expr, ...) ERR_IF(expr, -1 ==, ssize_t, ##__VA_ARGS__)
-#define ERR_MQ(expr, ...) ERR_IF(expr, (mqd_t) -1 ==, mqd_t, ##__VA_ARGS__)
-#define ERR_IN_ADDR_T(expr, ...) ERR_IF(expr, (in_addr_t) -1 ==, in_addr_t, ##__VA_ARGS__)
-#define ERR_BARRIER(expr, ...) ERR_IF(expr, PTHREAD_BARRIER_SERIAL_THREAD != exprVal && 0 !=, int, ##__VA_ARGS__)
-#define ERR_SEM(expr, ...) ERR_IF(expr, SEM_FAILED ==, sem_t, ##__VA_ARGS__)
-
-// versions that allow a single okay errno value:
-#define ERR_IF_(expr, condition, acceptableError, castType, ...) ERR_IF(expr, errno != acceptableError && condition, castType, ##__VA_ARGS__)
-#define ERR_NEG1_(expr, acceptableError, ...) ERR_IF_(expr, -1 ==, acceptableError, int, ##__VA_ARGS__)
-#define ERR_NEG_(expr, acceptableError, ...) ERR_IF_(expr, 0 >, acceptableError, int, ##__VA_ARGS__)
-#define ERR_NON_ZERO_(expr, acceptableError, ...) ERR_IF_(expr, 0 !=, acceptableError, int, ##__VA_ARGS__)
-#define ERR_EOF_(expr, acceptableError, ...) ERR_IF_(expr, EOF == , acceptableError, int, ##__VA_ARGS__)
-#define ERR_NULL_(expr, acceptableError, ...) ERR_IF_(expr, NULL ==, acceptableError, void*, ##__VA_ARGS__)
-#define ERR_SSIZE_T_(expr, acceptableError, ...) ERR_IF(expr, -1 ==, acceptableError, ssize_t, ##__VA_ARGS__)
-#define ERR_MQ_(expr, acceptableError, ...) ERR_IF_(expr, (mqd_t) -1 ==, acceptableError, mqd_t, ##__VA_ARGS__)
-#define ERR_IN_ADDR_T_(expr, acceptableError, ...) ERR_IF(expr, (in_addr_t) -1 ==, acceptableError, in_addr_t, ##__VA_ARGS__)
-#define ERR_BARRIER_(expr, ...) ERR_IF(expr, PTHREAD_BARRIER_SERIAL_THREAD != exprVal && 0 !=, acceptableError, int, ##__VA_ARGS__)
-#define ERR_SEM_(expr, acceptableError, ...) ERR_IF(expr, SEM_FAILED ==, acceptableError, sem_t, ##__VA_ARGS__)
 
 // version of DBGprintf that only compiles in DEBUG_MODE
 // kinda want DBGprintf to be nice and accesible normally
@@ -698,18 +129,6 @@ replaceBrace:
 	MIGHT_ERR_1_ARG(__VA_ARGS__),\
 	MIGHT_ERR_0_ARGS(__VA_ARGS__))
 
-#if EXTENSIONS_ALLOWED
-#define myRetry(expr, castType) (__extension__({\
-    castType val;\
-    do {val = (castType) (expr);} \
-    while(errno == EINTR);\
-    val;\
-}))
-#else // EXTENSIONS_ALLOWED
-// lol:
-#define myRetry(expr, castType) TEMP_FAILURE_RETRY(expr)
-#endif // EXTENSIONS_ALLOWED
-
 // So we use these to just fail if we're in RETURN_ERR
 // and the function fails? okayyy
 // we boutta need a big rework lol
@@ -720,21 +139,6 @@ replaceBrace:
 #define MIGHT_FAIL_NULL(expr, ...) MIGHT_FAIL_(expr, ERR_NULL, ##__VA_ARGS__)
 #define MIGHT_FAIL_SSIZE_T(expr, ...) MIGHT_FAIL_(expr, ERR_SSIZE_T, ##__VA_ARGS__)
 #define MIGHT_FAIL_IN_ADDR_T(expr, ...) MIGHT_FAIL_(expr, ERR_IN_ADDR_T, ##__VA_ARGS__)
-
-// UGHHH YUCK GROSSS
-// TODO: make these names better
-#define CHECK_RETRY_BASE(expr, err, errType, ...) err( myRetry(expr, errType) , ##__VA_ARGS__)
-
-#define CHECK_RETRY(expr, ...) CHECK_RETRY_BASE(expr, ERR_NEG1, int, ##__VA_ARGS__)
-#define CHECK_RETRY_MQ(expr, ...) CHECK_RETRY_BASE(expr, ERR_MQ, mqd_t, ##__VA_ARGS__)
-#define CHECK_RETRY_SSIZE_T(expr, ...) CHECK_RETRY_BASE(expr, ERR_SSIZE_T, ssize_t, ##__VA_ARGS__)
-#define CHECK_RETRY_EOF(expr, ...) CHECK_RETRY_BASE(expr, ERR_EOF, int, ##__VA_ARGS__)
-
-#define CHECK_RETRY_(expr, ...) CHECK_RETRY_BASE(expr, ERR_NEG1_, int, ##__VA_ARGS__)
-#define CHECK_RETRY_MQ_(expr, ...) CHECK_RETRY_BASE(expr, ERR_MQ_, mqd_t, ##__VA_ARGS__)
-#define CHECK_RETRY_SSIZE_T_(expr, ...) CHECK_RETRY_BASE(expr, ERR_SSIZE_T_, ssize_t, ##__VA_ARGS__)
-#define CHECK_RETRY_EOF_(expr, ...) CHECK_RETRY_BASE(expr, ERR_EOF_, int, ##__VA_ARGS__)
-// in theory there's more types to check but whoooooooo caaaaares
 
 // error checked and retried (if EINTR) versions of some message queue functions
 #define mq_send_(mqdes, msg_ptr, msg_len, msg_prio, ...)\
