+++ testing/makefile
@@ -1,4 +1,3 @@
-
 # base filename goes here:
 FILENAME=main
